+++ src/error.rs
@@ -1,0 +1,7 @@
+use thiserror::Error;
+
+#[derive(Debug, Error, Default)]
+pub enum Error {
+    #[default]
+    Unknown
+}
